LayoutTests/js/kde/Error-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Error('msg').message is 'msg'
PASS (new Error('msg')).message is 'msg'
PASS (new Error('msg')).name is 'Error'
PASS Object.prototype.toString.apply(Error()) is '[object Error]'
PASS Object.prototype.toString.apply(Error) is '[object Function]'
PASS Object.prototype.toString.apply(EvalError) is '[object Function]'
PASS successfullyParsed is true

TEST COMPLETE
